// src/api/schemas/getAgenciasAgencia200DataItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAC API
 * Api para uso interno WAC.
 * OpenAPI spec version: 1.0.16
 */

export type GetAgenciasAgencia200DataItem = {
  id?: string;
  email?: string;
  pais?: string;
  nombre?: string;
  telefono?: string;
  contacto?: string;
  comision?: string;
  /** @nullable */
  tipo_pago?: string | null;
  ts_creacion?: string;
  status?: string;
  fecha_creacion?: string;
  total_ventas?: string;
  total_ventas_monto?: string;
  total_comisiones?: string;
  cotizaciones_vigentes?: string;
  cotizaciones_vigentes_monto?: string;
};